requirements.txt
Pillow
alive-progress